+++ src/SparkPost.Tests/RequestSenders/RequestSenderTests.cs
@@ -3,8 +3,9 @@ using AutoMoq.Helpers;
 using Moq;
 using NUnit.Framework;
 using Should;
+using SparkPost.RequestSenders;
 
-namespace SparkPost.Tests
+namespace SparkPost.Tests.RequestSenders
 {
     public class RequestSenderTests
     {
