+++ src/SparkPost/Client.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Net.Http;
+using SparkPost.RequestSenders;
 
 namespace SparkPost
 {
