+++ src/SparkPost/RequestSenders/RequestSender.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace SparkPost
+namespace SparkPost.RequestSenders
 {
     public interface IRequestSender
     {
