+++ src/SparkPost/RequestSenders/SyncRequestSender.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 
-namespace SparkPost
+namespace SparkPost.RequestSenders
 {
     public class SyncRequestSender : IRequestSender
     {
